add generate_seeds script
Add a simple script to generate seeds with a number of adjectives and a noun. For
example the zpool name `practical-defiant-coffee`. This corpus could probably do
with more filtering, it seems to take quite a lot of goes to get a decent name.
These names are useful because they're human readable and memorable after a while,
unlike alternatives like random characters.