timewall: start daemon before exec_always set
If the `exec_always` that updates the wallpaper by name takes too long it appears to prevent the daemon from starting. Moving it so the `exec_always` is after the daemon flips this, so the exec always fails. Adding the wallpaper path to the daemon the first time around makes this okay - the rest of the time there is no collision.

This has exposed that the available timewall commands are imperfect. We want to start the daemon immediately and then update the wallpaper that _the daemon is using_ at will. The current status quo allows us to start the daemon naming the wallpaper, or switch the wallpaper, but not to update the wallpaper the daemon is using. After the geoclue changes to timewall this may need fixing. The current version makes the wallpaper disappear unnecessarily quite often, even if it would be putting the same wallpaper there anyway. One option is to only start the daemon and ignore wallpaper updates, or perhaps we can use a symlink to the wallpaper and update that instead. Will think about this more later.

Test Plan:
- Rebooted `merlin.rig` twice. Both times the daemon was running on starting sway.