nebula: disable nixos firewall globally
Previously I left it up to each host to allow full access on the Nebula interface. Changing that approach here to enabling this globally. This is because we should use the Nebula certificate capability based firewalling in future, but that is a big change that is a pain to roll-out so holding off on that for now.