gitea: cleanup old archives automatically
Getting my archives from the automatically rebasing nixpkgs/nixos-unstable generates and caches an archive. It appears these aren't getting cleaned up currently, and the directory is several hundred GB. Add the cron job in Gitea settings to automatically clean up archives older than 4 weeks. This should never affect my work flow as I update nixpkgs every 2 weeks usually, and if it is a miss it should be simple for the archive to be recreated.

Note: we should generate a dated tag in the nixos repo after each rebase to ensure the commit survives forever, as the rebase will otherwise lose the hash to garbage collection. It seems this was working before because the archives were kept forever.