nixpkgs-unstable: pull from github mirror instead of gitea
Pulling the archives from Gitea is causing a world of pain. Something seems to be downloading a tonne of different archives from the nixpkgs repo, and because they're all massive it's filling up the disk, which is breaking other services in painful ways. Rather than fixing the problem, mirror nixpkgs to GitHub and download from there instead. Kind of sad that this isn't fully self hosted anymore but we need GitHub to get ordinary nixpkgs anyway.