phoenix: start containers after zfs-mount
Even though Phoenix now boots successfully most of the time, these containers
always fail because they don't have a dependency on zfs-mount. If the entries
were in fstab I believe this would be implied as systemd does wait for specified
mounts, but they're not so the root mount looks like the right one. Specify
explicitly that these services require zfs-mount on this host and make them
start after it.