step-ca: fix impermanence mount
Modern systemd uses ID mapped mounts for DynamicUser support. This doesn't work if you have a mount below the state directory, including if that mount is a bind mount. This means that impermanence has to bind mount the StateDirectory with an all or nothing approach. Take this into account for the step-ca service.