restic: add aws mirror
Start the switch to AWS Deep Glacier for Restic storage. It's substantially
cheaper when you never restore, and given how unlikely a restore is with the
many other layers of backups this is fine.

For now, mirror just 128G to AWS in one region. Will expand to two regions and
all restic repos later. Also, don't forget anything yet. This storage is cheap
and we can work out the details of that later.