restic: sync 128G->1.6T
As we unify on single cloud repos with `restic copy` instead of multiple
copies, syncing 128G->1.6T becomes efficient. This means that hard
drives only have to sync a single repo instead of the current two which
makes the script more uniform, and won't result in duplicated snapshots
in the cloud. It also allows for smarter size based forgetting in the
future.