restic: rename repositories
The restic repository names were pretty bad and don't reflect their
usage well anymore. Specifically, 128G was named that way for the 128GB
flash drives it was cloned to, but these have been 256GB for years. 1.6T
was named that way for the 2TB hard drives it was cloned to because they
also had the 128GB repo, but this hasn't been accurate for a while now
either - specifically we restic copy all of the 128GB repo to 1.6T so we
only have to clone one repo to any storage device.

Update these names to better reflect their usage. Specifically:
- 128G -> mig29 - it's nimble and light, reflecting the fact the flash
  drives are more up to date (liveness) and that the repo must stay
  smaller.
- 1.6T -> b52 - it's heavy and slow, reflecting the less frequently
  synced hard drives that have way higher capacities.

Here's hoping this naming scheme is more flexible and fits future use
cases better.

These are actually new repos and not renames so that everything can have
`--copy-chunker-params` for better deduplication as copying between them
becomes more frequent. They have the same content as the previous repos
though.