merlin: keep amd gpu always active
Since activating Ollama and unplugging the HDMI from the discrete
graphics card merlin keeps getting into a sad state where the fans on
the GPU spin at 100% and it's not available in software.

Set this kernel parameter to keep the GPU from being fully disabled.
This might have something to do with hybrid GPU support, but as it's
often just an accelerator card and the kernel gets into a bad state
let's keep it awake.