devboxes: run `nix flake archive` for all repos
Devboxes are unable to evaluate flakes after a reboot due to
impermanence. This is very annoying for development repos, particularly
on bob which often starts with no Internet.

Add a service which runs `nix flake archive` for all branches of all
repositories and creates GC roots for them. This allows developing while
offline with existing local branches.